Export toolbar context in separate file
